Update percentages from uint256 to uint24

+++ contracts/exchange/interfaces/IExchangeFeePool.sol
@@ -4,24 +4,24 @@ pragma solidity ^0.8.0;
 interface IExchangeFeePool { 
 
     /******** View Functions ********/
-    function rate() external view returns(uint256);
+    function rate() external view returns(uint24);
     
     function totalFeePool(bytes4 _token) external view returns(uint256);
 
-    function distributionRates() external view returns(address[] memory _funds, uint256[] memory _percentages);
+    function distributionRates() external view returns(address[] memory _funds, uint24[] memory _percentages);
     
     /******** Mutative Functions ********/
-    function setRate(uint256 _rate) external;
+    function setRate(uint24 _rate) external;
     
-    function updateDistributionFunds(address[] memory _funds, uint256[] memory _percentages) external;
+    function updateDistributionFunds(address[] memory _funds, uint24[] memory _percentages) external;
 
     function depositRoyalty(bytes4 _token, address _tokenAddr, uint256 _amount) external;
 
     function distribute(bytes4 _token, address _tokenAddr) external;
     
     /*********************** Events *********************/
-    event FeeUpdated(address indexed operator, uint256 rate);
-    event FundsUpdated(address indexed operator, address[] funds, uint256[] percentages);
+    event FeeUpdated(address indexed operator, uint24 rate);
+    event FundsUpdated(address indexed operator, address[] funds, uint24[] percentages);
     event FundsDistributed(address indexed operator, address[] funds, uint256[] distributions);
     event ExchangeFeesPaid(bytes4 indexed token, address tokenAddr, uint256 amount);
 }
